added a new MainRisk class
